Implementasi penuh fitur reservasi pada inventory dan perbaikan relasi user.

+++ Inventory-Approval-App/database/migrations/2025_10_10_082256_update_lend_submissions_table_for_inventory_relation.php
@@ -10,9 +10,12 @@ return new class extends Migration
     {
         Schema::table('lend_submissions', function (Blueprint $table) {
             // Hapus kolom lama
-            $table->dropColumn('item_name');
+            if (Schema::hasColumn('lend_submissions', 'item_name')) {
+             $table->dropColumn('item_name');
+            }
             // Tambahkan kolom baru sebagai foreign key ke tabel inventories
-            $table->foreignId('inventory_id')->after('department')->constrained()->onDelete('cascade');
+            $table->foreignId('inventory_id')->nullable()->after('department')->constrained()->onDelete('cascade');
+            
         });
     }
